backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
PyJWT==2.9.0
pymongo==4.10.1
